code cleanup & some fixes
git-svn-id: svn://scm.gforge.inria.fr/svn/blobseer/trunk@88 43460812-3f57-0410-813e-9589cf9fb6d1

+++ provider/page_manager.hpp
@@ -10,7 +10,7 @@
 #include "common/cache_mt.hpp"
 #include "common/config.hpp"
 
-typedef boost::tuple<boost::int32_t, buffer_wrapper, boost::uint64_t, std::string> monitored_params_t;
+typedef boost::tuple<boost::uint64_t, buffer_wrapper, boost::uint64_t, std::string> monitored_params_t;
 
 template <class Persistency> class page_manager {
 private:
